PIM-4937: Fix LocaleListener and LocaleSubscriber
